Update eslint config
