+++ src/multiagent-system/modules/common/models/agentsEnvironment.ts
@@ -15,19 +15,13 @@ export default class AgentsEnvironment {
     private async ponderAgents(): Promise<PromiseSettledResult<void>[]> {
         const ponderPromises = this.agents.map(agent => {
             const nearestUnits = this.getNearestUnits(agent.unit);
-            const ponderPromise = agent.runPondering(nearestUnits);
-
-            return ponderPromise;
-        });
-
-        return Promise.allSettled(ponderPromises);
+            const ponderPromise = agent.runPondering(nearestUnits);            return ponderPromise;
+        });        return Promise.allSettled(ponderPromises);
     }
 
     private getNearestUnits(targetUnit: Unit): Array<Unit> {
         const nearestUnits = this.agents
             .map(agent => agent.unit)
-            .filter(unit => unit.isInRange(targetUnit));
-
-        return nearestUnits;
+            .filter(unit => unit.isInRange(targetUnit));        return nearestUnits;
     }
 }
