+++ src/multiagent-system/modules/common/models/distance.ts
