+++ src/multiagent-system/modules/common/models/figures/shape.ts
