+++ src/multiagent-system/modules/common/models/ponderingData.ts
