+++ src/multiagent-system/modules/common/models/speed.ts
