+++ src/multiagent-system/modules/melodyPlayer/models/note.ts
