+++ src/multiagent-system/modules/melodyPlayer/models/pianoKey.ts
